pyproject.toml
[tool.poetry]
name = "surfers camp utilities"
version = "0.1.0"
description = "A utility bot for the Surfers Camp discord server"
authors = ["arHSM <hanseungmin.ar@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
discord-py-interactions = "^4.2.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
pyright = "^1.1.254"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
